renamed scripts -> polling
